Skeleton for infinite.calendar jQuery library
Began the high-level layout of an infinitely scrollable calendaring
plugin for CalendarLanes. Currently triggers a callback when the user
scrolls a to within a configurable distance from the edge of a
scrollable region (default of 5%).

The callback causes the calendar to flash red right now, but it'll
eventually be responsible for pre-fetching data and rearranging calendar
panels (or something - not settled on the details yet) to give the
illusion of an infinitely scrollable module.